[core] Remove unnecessary code due changes in master in the meantime
